move registration to mainapp, change templates to use inheritance from base template
