+++ mainapp/templates/mainapp/index.html
@@ -1,80 +1,21 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <title>Aquatic Vehicles Registry</title>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    {% load static %}
-    <link rel="stylesheet" href="{% static 'css/bootstrap.min.css' %}">
-    <script src="{% static 'js/bootstrap.min.js' %}"></script>
-    <link rel="stylesheet" href="{% static 'css/mainapp.css' %}">
-</head>
-<body>
+{% extends "mainapp/base.html" %}
 
-<div class="test p-5 text-white text-center">
-    <div class="container">
-        <div class="row">
-            <div class="col-2">
-                <img src="{% static 'img/sailboat_icon_126071.ico' %}"/>
-            </div>
-            <div class="col-8">
-                <h1>Bureau of Aquatic Vehicles and Floating Devices</h1>
-                <p>Register your boats, surfboards and floaties here!</p>
-            </div>
-            <div class="col-2">
-            </div>
+{% block content %}
+    {% if user.is_authenticated %}
+        <div class="col-sm-12">
+            <h2>TITLE HEADING</h2>
+            <p>Some text..</p>
+            <p>Sunt in culpa qui officia deserunt mollit anim id est laborum consectetur adipiscing elit, sed do eiusmod
+                tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation
+                ullamco.</p>
+
+            <h2 class="mt-5">TITLE HEADING</h2>
+            <p>Some text..</p>
+            <p>Sunt in culpa qui officia deserunt mollit anim id est laborum consectetur adipiscing elit, sed do eiusmod
+                tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation
+                ullamco.</p>
         </div>
-    </div>
-</div>
-
-<nav class="topnavbar navbar navbar-expand-sm navbar-dark sticky-top">
-    <div class="container">
-        <ul class="navbar-nav">
-            <li class="nav-item">
-                <a class="nav-link active" href="#">Strona glowna</a>
-            </li>
-            <li class="nav-item">
-                <a class="nav-link" href="/formularz">Rejestracja pojazdu</a>
-            </li>
-            <li class="nav-item">
-                <a class="nav-link" href="#">Link</a>
-            </li>
-        </ul>
-        <ul class="nav navbar-nav navbar-right">
-            <li>
-                <div class="btn-nav"><a class="loginbutton btn btn-small navbar-btn" href="/login">Log in</a>
-                </div>
-            </li>
-        </ul>
-    </div>
-</nav>
-
-<div class="container mt-5">
-    <div class="row">
-        {% if user.is_authenticated %}
-            <div class="col-sm-12">
-                <h2>TITLE HEADING</h2>
-                <p>Some text..</p>
-                <p>Sunt in culpa qui officia deserunt mollit anim id est laborum consectetur adipiscing elit, sed do eiusmod
-                    tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation
-                    ullamco.</p>
-
-                <h2 class="mt-5">TITLE HEADING</h2>
-                <p>Some text..</p>
-                <p>Sunt in culpa qui officia deserunt mollit anim id est laborum consectetur adipiscing elit, sed do eiusmod
-                    tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation
-                    ullamco.</p>
-            </div>
-        {% else %}
-            <p> <a href="/login"> You have to login to access data</a></p>
-        {% endif %}
-
-    </div>
-</div>
-
-<div class="myfooter mt-5 p-2 text-white text-center fixed-bottom">
-    <p>Footer</p>
-</div>
-
-</body>
-</html>
+    {% else %}
+        <p> <a href="/login"> You have to login to access data</a></p>
+    {% endif %}
+{% endblock %}
