+++ mainapp/templates/registration/login.html
@@ -1,70 +1,10 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <title>Aquatic Vehicles Registry</title>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    {% load static %}
-    <link rel="stylesheet" href="{% static 'css/bootstrap.min.css' %}">
-    <script src="{% static 'js/bootstrap.min.js' %}"></script>
-    <link rel="stylesheet" href="{% static 'css/mainapp.css' %}">
-</head>
-<body>
+{% extends "mainapp/base.html" %}
 
-<div class="test p-5 text-white text-center">
-    <div class="container">
-        <div class="row">
-            <div class="col-2">
-                <img src="{% static 'img/sailboat_icon_126071.ico' %}"/>
-            </div>
-            <div class="col-8">
-                <h1>Bureau of Aquatic Vehicles and Floating Devices</h1>
-                <p>Register your boats, surfboards and floaties here!</p>
-            </div>
-            <div class="col-2">
-            </div>
-        </div>
-    </div>
-</div>
-
-<nav class="topnavbar navbar navbar-expand-sm navbar-dark sticky-top">
-    <div class="container">
-        <ul class="navbar-nav">
-            <li class="nav-item">
-                <a class="nav-link active" href="#">Strona glowna</a>
-            </li>
-            <li class="nav-item">
-                <a class="nav-link" href="/formularz">Rejestracja pojazdu</a>
-            </li>
-            <li class="nav-item">
-                <a class="nav-link" href="#">Link</a>
-            </li>
-        </ul>
-        <ul class="nav navbar-nav navbar-right">
-            <li>
-                <div class="btn-nav"><a class="loginbutton btn btn-small navbar-btn" href="/login">Log in</a>
-                </div>
-            </li>
-        </ul>
-    </div>
-</nav>
-
-<div class="container mt-5">
-    <div class="row">
-
-       <form method="post" class="from-group">
+{% block content %}
+    <form method="post" class="from-group">
         {%csrf_token %}
         {{form}}
-        <p>Don't have an account? Create one <a href="/register">here</a></p>
+        <p>Don't have an account? Create one <a href="/main/register">here</a></p>
         <button type="submit" class="btn btn-success">Login</button>
-        </form>
-    </div>
-</div>
-
-<div class="myfooter mt-5 p-2 text-white text-center fixed-bottom">
-    <p>Footer</p>
-</div>
-
-
-</body>
-</html>
+    </form>
+{% endblock %}
