+++ mainapp/templates/registration/register.html
@@ -1,0 +1,10 @@
+{% extends "mainapp/base.html" %}
+
+{% block content %}
+    {% load crispy_forms_tags %}
+       <form method="post" class="from-group">
+            {%csrf_token %}
+            {{form|crispy}}
+            <button type="submit" class="btn btn-success">asdasdasd</button>
+        </form>
+{% endblock %}
