+++ register/templates/base/petitioner_base.html
@@ -1,6 +1,4 @@
 <!DOCTYPE html>
-
-{% load crispy_forms_tags %}
 <html lang="en">
 <head>
     <title>Aquatic Vehicles Registry</title>
@@ -12,6 +10,7 @@
     <link rel="stylesheet" href="{% static 'css/mainapp.css' %}">
 </head>
 <body>
+
 <div class="test p-5 text-white text-center">
     <div class="container">
         <div class="row">
@@ -32,7 +31,7 @@
     <div class="container">
         <ul class="navbar-nav">
             <li class="nav-item">
-                <a class="nav-link active" href="/main">Strona glowna</a>
+                <a class="nav-link active" href="#">Strona glowna</a>
             </li>
             <li class="nav-item">
                 <a class="nav-link" href="/formularz">Rejestracja pojazdu</a>
@@ -52,19 +51,13 @@
 
 <div class="container mt-5">
     <div class="row">
-
-       <form method="post" class="from-group">
-            {%csrf_token %}
-            {{form|crispy}}
-            <button type="submit" class="btn btn-success">Submit</button>
-        </form>
+        {% block content %}{% endblock %}
     </div>
 </div>
 
 <div class="myfooter mt-5 p-2 text-white text-center fixed-bottom">
     <p>Footer</p>
 </div>
-
 
 </body>
 </html>
