+++ register/views.py
@@ -1,5 +1,4 @@
 from django.shortcuts import render, redirect
-from .forms import RegisterForm
 from .forms import VehRegisterForm
 
 # Create your views here.
@@ -26,13 +25,3 @@ def vehregister(request):
     context = {'form': form}
     return render(request, "formularze/vehregister.html", context)
 
-
-def register(response):
-    if response.method == "POST":
-        form = UserCreationForm(response.POST)
-        if form.is_valid():
-            form.save()
-        return redirect("/main")
-    else:
-        form = RegisterForm()
-    return render(response, "register/register.html", {"form": form})
