+++ shipregistry/urls.py
@@ -15,13 +15,11 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import include, path
-from register import views as v
 from register.views import vehregister
 
 urlpatterns = [
     path('main/', include('mainapp.urls')),
     path('admin/', admin.site.urls),
     path('', include("django.contrib.auth.urls")),
-    path("register/", v.register, name="register"),
     path("formularz/", vehregister),
 ]
